modified:   src/fits.f90

+++ src/fits.f90
@@ -36,7 +36,7 @@ module fits
 
         ! derived values
         character(len=16) :: flux = ''
-        real(kind=4) dmin, dmax
+        real(kind=c_float) dmin, dmax
         real(kind=4), allocatable :: frame_min(:), frame_max(:)
         real(kind=c_float), allocatable :: pixels(:, :)
         logical(kind=1), allocatable :: mask(:, :)
